requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-pinecone
langchain-chroma
langchain-upstage
langchain-ollama
supabase
pydantic
